+++ waku/v2/protocol/waku_filter/waku_filter_types.nim
@@ -10,7 +10,7 @@ export waku_message
 
 type
   ContentFilter* = object
-    contentTopics*: seq[ContentTopic]
+    contentTopic*: ContentTopic
 
   ContentFilterHandler* = proc(msg: WakuMessage) {.gcsafe, closure.}